Scraper_new/requirements.txt
nltk
sklearn
numpy
selenium
validator_collection
tldextract
screeninfo
gitpython
wget